fix(tsserver): verify that lsp.code is a string
In tsserver diagnostics, lsp.code is a number. Previously, the logic
used that number to match the value against "unused" or "never user",
which had no chance to match.

This commit fixes support for tsserver. It modifies the logic in a way
that the `lsp.code` is used as the message only if it is a string.

Fixes https://github.com/NarutoXY/dim.lua/issues/2