fix(hl_ns): use the global namespace for highlighting
Do not change Neovim's current highlight namespace to `dim.ns`. That
required setting the highlight groups in the `dim.ns` namespace. Aside
from the recent breaking change in Neovim's API for setting the `hl_ns`
[0], this means that this plugin expects no other plugin to manage the
Neovim's highlight namespace. The dim.lua plugin is not a highlight
manager, so IMO it should not set the Neovim highlight namespace.

Instead, the dim.lua highlight groups (e.g. `TSVariableDimmed` or
`TSParameterDimmed`) are defined in the **global** highlight namespace.
However, lines are highlighted using the `dim.ns` namespace. This makes
it so that `nvim_buf_clear_namespace` clears only the highlighted text
in the `dim.ns` namespace.

All in all, this change makes it so that the plugin does not assume
anything about the current Neovim highlight namespace and lets other
plugins set it, and it still works regardless, because it sets the
highlight groups in the global namespace.

Fixes https://github.com/NarutoXY/dim.lua/issues/15

[0]: https://github.com/neovim/neovim/commit/35653e6bcda6923b5213fb9356c35067d6d0288f